[Refactor]: 유지보수성을 위한 post와 comment controller 파일 분리
